+++ app/modules/__init__.py
@@ -2,7 +2,7 @@ from app.modules.admin import models as admin_models
 from app.modules.auth import models as auth_models
 from app.modules.community import models as community_models
 from app.modules.dayoff import models as dayoff_models
-from app.modules.mainpage import models as mainpage_models
+from app.modules.workstatus import models as workstatus_models
 from app.modules.payroll import models as payroll_models
 from app.modules.schedule import models as schedule_models
 from app.modules.shift import models as shift_models
@@ -17,5 +17,5 @@ __all__ = [
     "shift_models",
     "dayoff_models",
     "community_models",
-    "mainpage_models",
+    "workstatus_models",
 ]
