--- app/modules/mainpage/routers.py
@@ -1,8 +1,0 @@
-from fastapi import APIRouter
-
-router = APIRouter()
-
-
-@router.get("/")
-def get_mainpage():
-    return {"message": "Hello World"}
