+++ docker/docker-compose.dev.yml
@@ -30,7 +30,7 @@ services:
       MYSQL_PASSWORD: password
       TZ: Asia/Seoul
     ports:
-      - "3306:3306"
+      - "3307:3306"
     volumes:
       - db_data:/var/lib/mysql
     networks:
